Fix for unreliable IPv4 internet test in alpine-install.func
Old test only checked a single IP, which is not ideal if that IP is not up 100% of the time (and it has not been, historically).
Added two additional public resolver IPs to the test, e.g. 8.8.8.8 (Google) and 9.9.9.9 (CloudFlare).
If any one of the three IPs responds to a ping, the internet verification succeeds. This is much more reliable than using a single IP address.
Also moved down the conditional and else for better readability.